src/api/v1/views.py
from drf_spectacular.utils import extend_schema
from rest_framework import response, status, views

from api.v1.serializers import UserCreateSerializer
from api.v1.schemes import USER_CREATE_EXAMPLE, USER_CREATED_201


@extend_schema(
    request=UserCreateSerializer,
    summary="Регистрация пользователя",
    tags=["Users"],
    examples=[USER_CREATE_EXAMPLE],
    responses={
        status.HTTP_201_CREATED: USER_CREATED_201,
    },
)
class RegisrtyView(views.APIView):
    """
    Регистрация пользователей.
    """

    def post(self, request):
        serializer = UserCreateSerializer(data=request.data)
        if serializer.is_valid():
            serializer.save()
            return response.Response(
                serializer.data, status=status.HTTP_201_CREATED
            )
        return response.Response(
            serializer.errors, status=status.HTTP_400_BAD_REQUEST
        )
